Updates readme and setup file
This commit updates the README to include the details of the latest
version release and also updates the build badge to match the new
username that this project is now under.

Also, this commit updates the setup.py file and increments the bugfix
version. This makes the version of the application v1.1.1